+++ sandboxes/scripts/mocks/offers.py
