+++ sandboxes/scripts/mocks/recommendations.py
