+++ sandboxes/scripts/mocks/stocks.py
